Extract monster race monster part into the monster package
몬스터에 관련된 클래스들을 monster 패키지로 이동하였습니다.
