+++ README.md
@@ -104,3 +104,4 @@
 2. Menu 선택을 해야해서 Enum을 사용한 메뉴선택 방식을 사용하고자 했습니다.
 3. 메뉴 선택기능을 추가하였습니다.
 4. Input Output 을 담당하는 부분을 io 패키지로 이동하였습니다.
+5. Monster에 관련된 부분을 monster 패키지로 이동하였습니다.
